fixes, refactor
